+++ plugins/modules/application.py
@@ -32,7 +32,8 @@ options:
     description:
       - The state of the resource being configured by the module on the NetScaler
         ADC node.
-      - When C(absent) the resource will be deleted from the NetScaler ADC node.
+      - When C(absent), the resource will be deleted from the NetScaler ADC node.
+      - When C(imported), the resource will be imported on the NetScaler ADC node.
     type: str
   appname:
     type: str