+++ plugins/modules/cloudcredential.py
@@ -31,8 +31,8 @@ options:
     description:
       - The state of the resource being configured by the module on the NetScaler
         ADC node.
-      - When C(present) the resource will be created if needed and configured according
-        to the module's parameters.
+      - When C(present), the resource will be added/updated configured according to
+        the module's parameters.
     type: str
   applicationid:
     type: str