+++ plugins/modules/nsfeature.py
@@ -32,8 +32,8 @@ options:
     description:
       - The state of the resource being configured by the module on the NetScaler
         ADC node.
-      - When C(enabled) the resource will be enabled on the NetScaler ADC node.
-      - When C(disabled) the resource will be disabled on the NetScaler ADC node.
+      - When C(enabled), the resource will be enabled on the NetScaler ADC node.
+      - When C(disabled), the resource will be disabled on the NetScaler ADC node.
     type: str
   feature:
     type: list