+++ plugins/modules/sslfipskey.py
@@ -33,7 +33,9 @@ options:
     description:
       - The state of the resource being configured by the module on the NetScaler
         ADC node.
-      - When C(absent) the resource will be deleted from the NetScaler ADC node.
+      - When C(absent), the resource will be deleted from the NetScaler ADC node.
+      - When C(created), the `create` operation will be applied on the resource.
+      - When C(imported), the resource will be imported on the NetScaler ADC node.
     type: str
   curve:
     type: str