+++ plugins/modules/systementitydata.py
@@ -31,7 +31,7 @@ options:
     description:
       - The state of the resource being configured by the module on the NetScaler
         ADC node.
-      - When C(absent) the resource will be deleted from the NetScaler ADC node.
+      - When C(absent), the resource will be deleted from the NetScaler ADC node.
     type: str
   alldeleted:
     type: bool